requirements.txt
# Nova Apply - Requirements
# Python 3.11+

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web Scraping & Browser Automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# LLM Integration
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Data & Storage
# sqlite3 is built into Python 3.x - no package needed

# Email
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
